Refactor test utils into their own module
